+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [coming soon]
 
 - Correcting a bug in the list of license keys that was causing a table to be displayed as `block` and therefore it lost it's table properties.
+- Bug fix. Transient `hashToKey()` should return an un-prefixed key.
+- Bug fix. `dieInvalid()` referenced incorrectly.
 
 ## v160831.54041